service functions added
